a function that returns the number of bits you would need to flip to get from one number to another.

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -12,8 +12,8 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int result;
-	unsigned int count;
+	unsigned long int result;
+	unsigned long int count;
 
 	result = n ^ m;
 	for (count = 0; result > 0;)
